// src/public/js/enums/enums.ts
/**
 * Personal and other user profiles sections
 */
export enum SectionLiItems{
    ACCOUNT_SETTINGS = "li-account-settings", 
    FRIENDS = "li-friends", 
    INFORMATION = "li-information", 
    PHOTOS = "li-photos", 
    VIDEOS = "li-videos"
}
/**
 * Section to load when user clicks on a <li> item in left side panel
 */
export enum MyProfileSections{
    FRIENDS = "friends",
    INFORMATION = "information",
    PHOTOS = "photos",
    SETTINGS = "settings",
    VIDEOS = "videos"
}